Update source package format
