--- maas-schemas/lib/ValidationError.js
@@ -1,8 +1,0 @@
-'use strict';
-
-const { ValidationError } = require('maasglobal-json-schema-validator');
-
-/**
- * @deprecated import { ValidationError } from 'maasglobal-json-schema-validator'
- */
-module.exports = ValidationError;
