+++ maas-schemas/lib/ajv/registry.d.ts
@@ -1,0 +1,1 @@
+export default any;
