+++ maas-schemas/schemas.json
@@ -1,0 +1,4 @@
+{
+  "base": "https://schemas.maas.global/",
+  "deps": {}
+}
