+++ maas-schemas/test/feature-require-schemas.js
@@ -1,9 +1,10 @@
 'use strict';
 
+const path = require('path');
+
 const { testSchemaPackage } = require('maasglobal-json-schema-validator/lib/conformance');
 
 describe('Conformance tests', () => {
-  const packageRoot = '.';
-  const baseURI = 'https://schemas.maas.global/';
-  testSchemaPackage(packageRoot, baseURI, []);
+  const packageRoot = path.resolve(__dirname, '../../maas-schemas');
+  testSchemaPackage(packageRoot);
 });
