--- registry.js
@@ -1,251 +1,0 @@
-'use strict';
-module.exports = {
-  'schemas/core/booking-meta.json': require('./schemas/core/booking-meta.json'),
-  'schemas/core/booking-option.json': require('./schemas/core/booking-option.json'),
-  'schemas/core/booking.json': require('./schemas/core/booking.json'),
-  'schemas/core/card.json': require('./schemas/core/card.json'),
-  'schemas/core/customer.json': require('./schemas/core/customer.json'),
-  'schemas/core/error.json': require('./schemas/core/error.json'),
-  'schemas/core/iot-thing-shadow.json': require('./schemas/core/iot-thing-shadow.json'),
-  'schemas/core/itinerary.json': require('./schemas/core/itinerary.json'),
-  'schemas/core/leg.json': require('./schemas/core/leg.json'),
-  'schemas/core/paymentSource.json': require('./schemas/core/paymentSource.json'),
-  'schemas/core/plan.json': require('./schemas/core/plan.json'),
-  'schemas/core/product-option.json': require('./schemas/core/product-option.json'),
-  'schemas/core/product.json': require('./schemas/core/product.json'),
-  'schemas/core/profile.json': require('./schemas/core/profile.json'),
-  'schemas/core/region.json': require('./schemas/core/region.json'),
-  'schemas/core/stop.json': require('./schemas/core/stop.json'),
-  'schemas/geojson/geometry.json': require('./schemas/geojson/geometry.json'),
-  'schemas/core/components/ACRISS.json': require('./schemas/core/components/ACRISS.json'),
-  'schemas/core/components/address.json': require('./schemas/core/components/address.json'),
-  'schemas/core/components/agencyOptions.json': require('./schemas/core/components/agencyOptions.json'),
-  'schemas/core/components/api-common.json': require('./schemas/core/components/api-common.json'),
-  'schemas/core/components/authorization.json': require('./schemas/core/components/authorization.json'),
-  'schemas/core/components/bike-station.json': require('./schemas/core/components/bike-station.json'),
-  'schemas/core/components/car-rental.json': require('./schemas/core/components/car-rental.json'),
-  'schemas/core/components/common.json': require('./schemas/core/components/common.json'),
-  'schemas/core/components/configurator.json': require('./schemas/core/components/configurator.json'),
-  'schemas/core/components/cost.json': require('./schemas/core/components/cost.json'),
-  'schemas/core/components/costBreakDown.json': require('./schemas/core/components/costBreakDown.json'),
-  'schemas/core/components/customerSelection.json': require('./schemas/core/components/customerSelection.json'),
-  'schemas/core/components/fare.json': require('./schemas/core/components/fare.json'),
-  'schemas/core/components/geolocation.json': require('./schemas/core/components/geolocation.json'),
-  'schemas/core/components/geometry.json': require('./schemas/core/components/geometry.json'),
-  'schemas/core/components/i18n.json': require('./schemas/core/components/i18n.json'),
-  'schemas/core/components/message.json': require('./schemas/core/components/message.json'),
-  'schemas/core/components/payment-parameters.json': require('./schemas/core/components/payment-parameters.json'),
-  'schemas/core/components/personalDataAllowItem.json': require('./schemas/core/components/personalDataAllowItem.json'),
-  'schemas/core/components/place.json': require('./schemas/core/components/place.json'),
-  'schemas/core/components/point-cost.json': require('./schemas/core/components/point-cost.json'),
-  'schemas/core/components/spaceDemand.json': require('./schemas/core/components/spaceDemand.json'),
-  'schemas/core/components/state-log.json': require('./schemas/core/components/state-log.json'),
-  'schemas/core/components/state.json': require('./schemas/core/components/state.json'),
-  'schemas/core/components/station.json': require('./schemas/core/components/station.json'),
-  'schemas/core/components/subscriptionChangeState.json': require('./schemas/core/components/subscriptionChangeState.json'),
-  'schemas/core/components/terms.json': require('./schemas/core/components/terms.json'),
-  'schemas/core/components/travel-mode.json': require('./schemas/core/components/travel-mode.json'),
-  'schemas/core/components/units-geo.json': require('./schemas/core/components/units-geo.json'),
-  'schemas/core/components/units.json': require('./schemas/core/components/units.json'),
-  'schemas/core/modes/MODE_BICYCLE.json': require('./schemas/core/modes/MODE_BICYCLE.json'),
-  'schemas/core/modes/MODE_BUS.json': require('./schemas/core/modes/MODE_BUS.json'),
-  'schemas/core/modes/MODE_BUSISH.json': require('./schemas/core/modes/MODE_BUSISH.json'),
-  'schemas/core/modes/MODE_CABLE_CAR.json': require('./schemas/core/modes/MODE_CABLE_CAR.json'),
-  'schemas/core/modes/MODE_CAR.json': require('./schemas/core/modes/MODE_CAR.json'),
-  'schemas/core/modes/MODE_FERRY.json': require('./schemas/core/modes/MODE_FERRY.json'),
-  'schemas/core/modes/MODE_FUNICULAR.json': require('./schemas/core/modes/MODE_FUNICULAR.json'),
-  'schemas/core/modes/MODE_GONDOLA.json': require('./schemas/core/modes/MODE_GONDOLA.json'),
-  'schemas/core/modes/MODE_RAIL.json': require('./schemas/core/modes/MODE_RAIL.json'),
-  'schemas/core/modes/MODE_SCOOTER.json': require('./schemas/core/modes/MODE_SCOOTER.json'),
-  'schemas/core/modes/MODE_SHARED_BICYCLE.json': require('./schemas/core/modes/MODE_SHARED_BICYCLE.json'),
-  'schemas/core/modes/MODE_SHARED_CAR.json': require('./schemas/core/modes/MODE_SHARED_CAR.json'),
-  'schemas/core/modes/MODE_SUBWAY.json': require('./schemas/core/modes/MODE_SUBWAY.json'),
-  'schemas/core/modes/MODE_TAXI.json': require('./schemas/core/modes/MODE_TAXI.json'),
-  'schemas/core/modes/MODE_TRAIN.json': require('./schemas/core/modes/MODE_TRAIN.json'),
-  'schemas/core/modes/MODE_TRAINISH.json': require('./schemas/core/modes/MODE_TRAINISH.json'),
-  'schemas/core/modes/MODE_TRAM.json': require('./schemas/core/modes/MODE_TRAM.json'),
-  'schemas/core/modes/MODE_TRANSIT.json': require('./schemas/core/modes/MODE_TRANSIT.json'),
-  'schemas/core/modes/MODE_WALK.json': require('./schemas/core/modes/MODE_WALK.json'),
-  'schemas/maas-backend/booking-option-create/request.json': require('./schemas/maas-backend/booking-option-create/request.json'),
-  'schemas/maas-backend/booking-option-create/response.json': require('./schemas/maas-backend/booking-option-create/response.json'),
-  'schemas/maas-backend/customers/customer.json': require('./schemas/maas-backend/customers/customer.json'),
-  'schemas/maas-backend/customers/personalData.json': require('./schemas/maas-backend/customers/personalData.json'),
-  'schemas/maas-backend/invoices/invoice.json': require('./schemas/maas-backend/invoices/invoice.json'),
-  'schemas/maas-backend/invoices/invoiceLineItem.json': require('./schemas/maas-backend/invoices/invoiceLineItem.json'),
-  'schemas/maas-backend/invoices/invoiceUnits.json': require('./schemas/maas-backend/invoices/invoiceUnits.json'),
-  'schemas/maas-backend/products/provider.json': require('./schemas/maas-backend/products/provider.json'),
-  'schemas/maas-backend/push-notification/request.json': require('./schemas/maas-backend/push-notification/request.json'),
-  'schemas/maas-backend/push-notification/response.json': require('./schemas/maas-backend/push-notification/response.json'),
-  'schemas/maas-backend/subscriptions/contact.json': require('./schemas/maas-backend/subscriptions/contact.json'),
-  'schemas/maas-backend/subscriptions/pricing.json': require('./schemas/maas-backend/subscriptions/pricing.json'),
-  'schemas/maas-backend/subscriptions/subscription.json': require('./schemas/maas-backend/subscriptions/subscription.json'),
-  'schemas/maas-backend/subscriptions/subscriptionAddress.json': require('./schemas/maas-backend/subscriptions/subscriptionAddress.json'),
-  'schemas/maas-backend/subscriptions/subscriptionOption.json': require('./schemas/maas-backend/subscriptions/subscriptionOption.json'),
-  'schemas/tsp/booking-cancel/request.json': require('./schemas/tsp/booking-cancel/request.json'),
-  'schemas/tsp/booking-cancel/response.json': require('./schemas/tsp/booking-cancel/response.json'),
-  'schemas/tsp/booking-create/request.json': require('./schemas/tsp/booking-create/request.json'),
-  'schemas/tsp/booking-create/response.json': require('./schemas/tsp/booking-create/response.json'),
-  'schemas/tsp/booking-options-list/request.json': require('./schemas/tsp/booking-options-list/request.json'),
-  'schemas/tsp/booking-options-list/response.json': require('./schemas/tsp/booking-options-list/response.json'),
-  'schemas/tsp/booking-read-by-id/request.json': require('./schemas/tsp/booking-read-by-id/request.json'),
-  'schemas/tsp/booking-read-by-id/response.json': require('./schemas/tsp/booking-read-by-id/response.json'),
-  'schemas/tsp/booking-receipt/request.json': require('./schemas/tsp/booking-receipt/request.json'),
-  'schemas/tsp/booking-receipt/response.json': require('./schemas/tsp/booking-receipt/response.json'),
-  'schemas/tsp/booking-ticket/request.json': require('./schemas/tsp/booking-ticket/request.json'),
-  'schemas/tsp/booking-ticket/response.json': require('./schemas/tsp/booking-ticket/response.json'),
-  'schemas/tsp/booking-update/request.json': require('./schemas/tsp/booking-update/request.json'),
-  'schemas/tsp/booking-update/response.json': require('./schemas/tsp/booking-update/response.json'),
-  'schemas/tsp/customer-auth/request.json': require('./schemas/tsp/customer-auth/request.json'),
-  'schemas/tsp/customer-auth/response.json': require('./schemas/tsp/customer-auth/response.json'),
-  'schemas/tsp/customer-auth-validate/request.json': require('./schemas/tsp/customer-auth-validate/request.json'),
-  'schemas/tsp/customer-auth-validate/response.json': require('./schemas/tsp/customer-auth-validate/response.json'),
-  'schemas/tsp/journey-planner/request.json': require('./schemas/tsp/journey-planner/request.json'),
-  'schemas/tsp/journey-planner/response.json': require('./schemas/tsp/journey-planner/response.json'),
-  'schemas/tsp/stations-list/request.json': require('./schemas/tsp/stations-list/request.json'),
-  'schemas/tsp/stations-list/response.json': require('./schemas/tsp/stations-list/response.json'),
-  'schemas/tsp/stations-retrieve/request.json': require('./schemas/tsp/stations-retrieve/request.json'),
-  'schemas/tsp/stations-retrieve/response.json': require('./schemas/tsp/stations-retrieve/response.json'),
-  'schemas/tsp/vehicle-alert/request.json': require('./schemas/tsp/vehicle-alert/request.json'),
-  'schemas/tsp/vehicle-alert/response.json': require('./schemas/tsp/vehicle-alert/response.json'),
-  'schemas/tsp/webhooks-bookings-update/remote-request.json': require('./schemas/tsp/webhooks-bookings-update/remote-request.json'),
-  'schemas/tsp/webhooks-bookings-update/remote-response.json': require('./schemas/tsp/webhooks-bookings-update/remote-response.json'),
-  'schemas/maas-backend/auth/auth-sms-login/request.json': require('./schemas/maas-backend/auth/auth-sms-login/request.json'),
-  'schemas/maas-backend/auth/auth-sms-request-code/request.json': require('./schemas/maas-backend/auth/auth-sms-request-code/request.json'),
-  'schemas/maas-backend/autocomplete/autocomplete-query/request.json': require('./schemas/maas-backend/autocomplete/autocomplete-query/request.json'),
-  'schemas/maas-backend/autocomplete/autocomplete-query/response.json': require('./schemas/maas-backend/autocomplete/autocomplete-query/response.json'),
-  'schemas/maas-backend/bookings/bookings-agency-options/request.json': require('./schemas/maas-backend/bookings/bookings-agency-options/request.json'),
-  'schemas/maas-backend/bookings/bookings-agency-options/response.json': require('./schemas/maas-backend/bookings/bookings-agency-options/response.json'),
-  'schemas/maas-backend/bookings/bookings-agency-products/request.json': require('./schemas/maas-backend/bookings/bookings-agency-products/request.json'),
-  'schemas/maas-backend/bookings/bookings-agency-products/response.json': require('./schemas/maas-backend/bookings/bookings-agency-products/response.json'),
-  'schemas/maas-backend/bookings/bookings-cancel/request.json': require('./schemas/maas-backend/bookings/bookings-cancel/request.json'),
-  'schemas/maas-backend/bookings/bookings-cancel/response.json': require('./schemas/maas-backend/bookings/bookings-cancel/response.json'),
-  'schemas/maas-backend/bookings/bookings-create/request.json': require('./schemas/maas-backend/bookings/bookings-create/request.json'),
-  'schemas/maas-backend/bookings/bookings-create/response.json': require('./schemas/maas-backend/bookings/bookings-create/response.json'),
-  'schemas/maas-backend/bookings/bookings-list/request.json': require('./schemas/maas-backend/bookings/bookings-list/request.json'),
-  'schemas/maas-backend/bookings/bookings-list/response.json': require('./schemas/maas-backend/bookings/bookings-list/response.json'),
-  'schemas/maas-backend/bookings/bookings-retrieve/request.json': require('./schemas/maas-backend/bookings/bookings-retrieve/request.json'),
-  'schemas/maas-backend/bookings/bookings-retrieve/response.json': require('./schemas/maas-backend/bookings/bookings-retrieve/response.json'),
-  'schemas/maas-backend/bookings/bookings-update/request.json': require('./schemas/maas-backend/bookings/bookings-update/request.json'),
-  'schemas/maas-backend/bookings/bookings-update/response.json': require('./schemas/maas-backend/bookings/bookings-update/response.json'),
-  'schemas/maas-backend/customers/payment-sources/paymentSource.json': require('./schemas/maas-backend/customers/payment-sources/paymentSource.json'),
-  'schemas/maas-backend/customers/personal-data-delete/request.json': require('./schemas/maas-backend/customers/personal-data-delete/request.json'),
-  'schemas/maas-backend/customers/retrieve/request.json': require('./schemas/maas-backend/customers/retrieve/request.json'),
-  'schemas/maas-backend/customers/retrieve/response.json': require('./schemas/maas-backend/customers/retrieve/response.json'),
-  'schemas/maas-backend/customers/stats/request.json': require('./schemas/maas-backend/customers/stats/request.json'),
-  'schemas/maas-backend/customers/stats/response.json': require('./schemas/maas-backend/customers/stats/response.json'),
-  'schemas/maas-backend/customers/update/request.json': require('./schemas/maas-backend/customers/update/request.json'),
-  'schemas/maas-backend/customers/update/response.json': require('./schemas/maas-backend/customers/update/response.json'),
-  'schemas/maas-backend/customers/benefits/initiate/request.json': require('./schemas/maas-backend/customers/benefits/initiate/request.json'),
-  'schemas/maas-backend/customers/benefits/initiate/response.json': require('./schemas/maas-backend/customers/benefits/initiate/response.json'),
-  'schemas/maas-backend/customers/benefits/callback/request.json': require('./schemas/maas-backend/customers/benefits/callback/request.json'),
-  'schemas/maas-backend/geocoding/geocoding-query/request.json': require('./schemas/maas-backend/geocoding/geocoding-query/request.json'),
-  'schemas/maas-backend/geocoding/geocoding-query/response.json': require('./schemas/maas-backend/geocoding/geocoding-query/response.json'),
-  'schemas/maas-backend/geocoding/geocoding-reverse/request.json': require('./schemas/maas-backend/geocoding/geocoding-reverse/request.json'),
-  'schemas/maas-backend/geocoding/geocoding-reverse/response.json': require('./schemas/maas-backend/geocoding/geocoding-reverse/response.json'),
-  'schemas/maas-backend/itineraries/itinerary-cancel/request.json': require('./schemas/maas-backend/itineraries/itinerary-cancel/request.json'),
-  'schemas/maas-backend/itineraries/itinerary-cancel/response.json': require('./schemas/maas-backend/itineraries/itinerary-cancel/response.json'),
-  'schemas/maas-backend/itineraries/itinerary-create/request.json': require('./schemas/maas-backend/itineraries/itinerary-create/request.json'),
-  'schemas/maas-backend/itineraries/itinerary-create/response.json': require('./schemas/maas-backend/itineraries/itinerary-create/response.json'),
-  'schemas/maas-backend/itineraries/itinerary-list/request.json': require('./schemas/maas-backend/itineraries/itinerary-list/request.json'),
-  'schemas/maas-backend/itineraries/itinerary-list/response.json': require('./schemas/maas-backend/itineraries/itinerary-list/response.json'),
-  'schemas/maas-backend/itineraries/itinerary-retrieve/request.json': require('./schemas/maas-backend/itineraries/itinerary-retrieve/request.json'),
-  'schemas/maas-backend/itineraries/itinerary-retrieve/response.json': require('./schemas/maas-backend/itineraries/itinerary-retrieve/response.json'),
-  'schemas/maas-backend/itineraries/itinerary-update/request.json': require('./schemas/maas-backend/itineraries/itinerary-update/request.json'),
-  'schemas/maas-backend/itineraries/itinerary-update/response.json': require('./schemas/maas-backend/itineraries/itinerary-update/response.json'),
-  'schemas/maas-backend/products/products-providers-options/request.json': require('./schemas/maas-backend/products/products-providers-options/request.json'),
-  'schemas/maas-backend/products/products-providers-options/response.json': require('./schemas/maas-backend/products/products-providers-options/response.json'),
-  'schemas/maas-backend/products/products-providers-retrieve/request.json': require('./schemas/maas-backend/products/products-providers-retrieve/request.json'),
-  'schemas/maas-backend/products/products-providers-retrieve/response.json': require('./schemas/maas-backend/products/products-providers-retrieve/response.json'),
-  'schemas/maas-backend/profile/profile-devices-put/request.json': require('./schemas/maas-backend/profile/profile-devices-put/request.json'),
-  'schemas/maas-backend/profile/profile-devices-put/response.json': require('./schemas/maas-backend/profile/profile-devices-put/response.json'),
-  'schemas/maas-backend/profile/profile-edit/request.json': require('./schemas/maas-backend/profile/profile-edit/request.json'),
-  'schemas/maas-backend/profile/profile-edit/response.json': require('./schemas/maas-backend/profile/profile-edit/response.json'),
-  'schemas/maas-backend/profile/profile-favoriteLocations-add/request.json': require('./schemas/maas-backend/profile/profile-favoriteLocations-add/request.json'),
-  'schemas/maas-backend/profile/profile-favoriteLocations-add/response.json': require('./schemas/maas-backend/profile/profile-favoriteLocations-add/response.json'),
-  'schemas/maas-backend/profile/profile-favoriteLocations-delete/request.json': require('./schemas/maas-backend/profile/profile-favoriteLocations-delete/request.json'),
-  'schemas/maas-backend/profile/profile-favoriteLocations-delete/response.json': require('./schemas/maas-backend/profile/profile-favoriteLocations-delete/response.json'),
-  'schemas/maas-backend/profile/profile-info/request.json': require('./schemas/maas-backend/profile/profile-info/request.json'),
-  'schemas/maas-backend/profile/profile-info/response.json': require('./schemas/maas-backend/profile/profile-info/response.json'),
-  'schemas/maas-backend/profile/profile-webhook/request.json': require('./schemas/maas-backend/profile/profile-webhook/request.json'),
-  'schemas/maas-backend/profile/profile-webhook/response.json': require('./schemas/maas-backend/profile/profile-webhook/response.json'),
-  'schemas/maas-backend/provider/autocomplete/request.json': require('./schemas/maas-backend/provider/autocomplete/request.json'),
-  'schemas/maas-backend/provider/autocomplete/response.json': require('./schemas/maas-backend/provider/autocomplete/response.json'),
-  'schemas/maas-backend/provider/geocoding/request.json': require('./schemas/maas-backend/provider/geocoding/request.json'),
-  'schemas/maas-backend/provider/geocoding/response.json': require('./schemas/maas-backend/provider/geocoding/response.json'),
-  'schemas/maas-backend/provider/geocoding-reverse/request.json': require('./schemas/maas-backend/provider/geocoding-reverse/request.json'),
-  'schemas/maas-backend/provider/geocoding-reverse/response.json': require('./schemas/maas-backend/provider/geocoding-reverse/response.json'),
-  'schemas/maas-backend/provider/routes/request.json': require('./schemas/maas-backend/provider/routes/request.json'),
-  'schemas/maas-backend/provider/routes/response.json': require('./schemas/maas-backend/provider/routes/response.json'),
-  'schemas/maas-backend/provider/sms/request.json': require('./schemas/maas-backend/provider/sms/request.json'),
-  'schemas/maas-backend/provider/sms/response.json': require('./schemas/maas-backend/provider/sms/response.json'),
-  'schemas/maas-backend/regions/regions-options/request.json': require('./schemas/maas-backend/regions/regions-options/request.json'),
-  'schemas/maas-backend/regions/regions-options/response.json': require('./schemas/maas-backend/regions/regions-options/response.json'),
-  'schemas/maas-backend/routes/routes-query/request.json': require('./schemas/maas-backend/routes/routes-query/request.json'),
-  'schemas/maas-backend/routes/routes-query/response.json': require('./schemas/maas-backend/routes/routes-query/response.json'),
-  'schemas/maas-backend/stations/stations-list/request.json': require('./schemas/maas-backend/stations/stations-list/request.json'),
-  'schemas/maas-backend/stations/stations-list/response.json': require('./schemas/maas-backend/stations/stations-list/response.json'),
-  'schemas/maas-backend/stations/stations-retrieve/request.json': require('./schemas/maas-backend/stations/stations-retrieve/request.json'),
-  'schemas/maas-backend/stations/stations-retrieve/response.json': require('./schemas/maas-backend/stations/stations-retrieve/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-change-state/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-change-state/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-change-state/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-change-state/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-create/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-create/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-create/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-create/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-customer-retrieve/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-customer-retrieve/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-customer-retrieve/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-customer-retrieve/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-customer-update/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-customer-update/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-customer-update/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-customer-update/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-estimate/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-estimate/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-estimate/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-estimate/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-create/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-create/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-create/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-create/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-update/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-update/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-update/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-update/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-list/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-list/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-list/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-list/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-retrieve/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-retrieve/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-retrieve/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-retrieve/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-update-state/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-update-state/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-intents-update-state/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-intents-update-state/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-options/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-options/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-options/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-options/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-retrieve/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-retrieve/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-retrieve/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-retrieve/response.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-update/request.json': require('./schemas/maas-backend/subscriptions/subscriptions-update/request.json'),
-  'schemas/maas-backend/subscriptions/subscriptions-update/response.json': require('./schemas/maas-backend/subscriptions/subscriptions-update/response.json'),
-  'schemas/maas-backend/tracking/tracking-active-itinerary-get/request.json': require('./schemas/maas-backend/tracking/tracking-active-itinerary-get/request.json'),
-  'schemas/maas-backend/tracking/tracking-active-itinerary-get/response.json': require('./schemas/maas-backend/tracking/tracking-active-itinerary-get/response.json'),
-  'schemas/maas-backend/tracking/tracking-active-itinerary-set/request.json': require('./schemas/maas-backend/tracking/tracking-active-itinerary-set/request.json'),
-  'schemas/maas-backend/tracking/tracking-active-itinerary-set/response.json': require('./schemas/maas-backend/tracking/tracking-active-itinerary-set/response.json'),
-  'schemas/maas-backend/tracking/tracking-active-leg-get/request.json': require('./schemas/maas-backend/tracking/tracking-active-leg-get/request.json'),
-  'schemas/maas-backend/tracking/tracking-active-leg-get/response.json': require('./schemas/maas-backend/tracking/tracking-active-leg-get/response.json'),
-  'schemas/maas-backend/tracking/tracking-active-leg-set/request.json': require('./schemas/maas-backend/tracking/tracking-active-leg-set/request.json'),
-  'schemas/maas-backend/tracking/tracking-active-leg-set/response.json': require('./schemas/maas-backend/tracking/tracking-active-leg-set/response.json'),
-  'schemas/maas-backend/tracking/tracking-cancel-active-itinerary/request.json': require('./schemas/maas-backend/tracking/tracking-cancel-active-itinerary/request.json'),
-  'schemas/maas-backend/tracking/tracking-cancel-active-itinerary/response.json': require('./schemas/maas-backend/tracking/tracking-cancel-active-itinerary/response.json'),
-  'schemas/maas-backend/tracking/tracking-update-user-location/request.json': require('./schemas/maas-backend/tracking/tracking-update-user-location/request.json'),
-  'schemas/maas-backend/tracking/tracking-update-user-location/response.json': require('./schemas/maas-backend/tracking/tracking-update-user-location/response.json'),
-  'schemas/maas-backend/tsp-auth/expire/request.json': require('./schemas/maas-backend/tsp-auth/expire/request.json'),
-  'schemas/maas-backend/tsp-auth/init/request.json': require('./schemas/maas-backend/tsp-auth/init/request.json'),
-  'schemas/maas-backend/tsp-auth/init/response.json': require('./schemas/maas-backend/tsp-auth/init/response.json'),
-  'schemas/maas-backend/tsp-auth/set/request.json': require('./schemas/maas-backend/tsp-auth/set/request.json'),
-  'schemas/maas-backend/tsp-auth/validate/request.json': require('./schemas/maas-backend/tsp-auth/validate/request.json'),
-  'schemas/maas-backend/tsp-auth/validate/response.json': require('./schemas/maas-backend/tsp-auth/validate/response.json'),
-  'schemas/maas-backend/vehicle/vehicle-alert/request.json': require('./schemas/maas-backend/vehicle/vehicle-alert/request.json'),
-  'schemas/maas-backend/vehicle/vehicle-alert/response.json': require('./schemas/maas-backend/vehicle/vehicle-alert/response.json'),
-  'schemas/maas-backend/webhooks/webhooks-bookings-update/request.json': require('./schemas/maas-backend/webhooks/webhooks-bookings-update/request.json'),
-  'schemas/maas-backend/webhooks/webhooks-bookings-update/response.json': require('./schemas/maas-backend/webhooks/webhooks-bookings-update/response.json'),
-  'schemas/maas-backend/webhooks/webhooks-payments/request.json': require('./schemas/maas-backend/webhooks/webhooks-payments/request.json'),
-  'schemas/maas-backend/webhooks/webhooks-payments/response.json': require('./schemas/maas-backend/webhooks/webhooks-payments/response.json'),
-  'schemas/maas-backend/webhooks/zendesk-push-notification/request.json': require('./schemas/maas-backend/webhooks/zendesk-push-notification/request.json'),
-  'schemas/maas-backend/webhooks/zendesk-push-notification/response.json': require('./schemas/maas-backend/webhooks/zendesk-push-notification/response.json'),
-  'schemas/maas-backend/bookings/v2/bookings-create/request.json': require('./schemas/maas-backend/bookings/v2/bookings-create/request.json'),
-  'schemas/maas-backend/bookings/v2/bookings-create/response.json': require('./schemas/maas-backend/bookings/v2/bookings-create/response.json'),
-  'schemas/maas-backend/customers/payment-sources/create/request.json': require('./schemas/maas-backend/customers/payment-sources/create/request.json'),
-  'schemas/maas-backend/customers/payment-sources/delete/request.json': require('./schemas/maas-backend/customers/payment-sources/delete/request.json'),
-  'schemas/maas-backend/customers/payment-sources/setup-intent/request.json': require('./schemas/maas-backend/customers/payment-sources/setup-intent/request.json'),
-  'schemas/maas-backend/customers/payment-sources/setup-intent/response.json': require('./schemas/maas-backend/customers/payment-sources/setup-intent/response.json'),
-  'schemas/maas-backend/customers/payment-sources/update/request.json': require('./schemas/maas-backend/customers/payment-sources/update/request.json'),
-  'schemas/maas-backend/customers/payment-sources/update/response.json': require('./schemas/maas-backend/customers/payment-sources/update/response.json'),
-  'schemas/maas-backend/webhooks/webhooks-payments/gateway/avainpay.json': require('./schemas/maas-backend/webhooks/webhooks-payments/gateway/avainpay.json'),
-  'schemas/maas-backend/webhooks/webhooks-payments/gateway/stripe.json': require('./schemas/maas-backend/webhooks/webhooks-payments/gateway/stripe.json'),
-};
